+++ src/layer/arm/gemm_int8_bf16s.h
@@ -174,7 +174,7 @@ static void compute_A_tile_bf16_int8_scales(const Mat& A, Mat& scales, float B_s
 #endif // __ARM_NEON
             for (; kk < K; kk++)
             {
-                absmax = std::max(absmax, (float)fabs(bfloat16_to_float32(p0[0])));
+                absmax = std::max(absmax, (float)fabsf(bfloat16_to_float32(p0[0])));
                 p0++;
             }
 
@@ -1319,7 +1319,7 @@ static void transpose_compute_A_tile_bf16_int8_scales(const Mat& A, Mat& scales,
             float absmax = 0.f;
             for (int kk = 0; kk < K; kk++)
             {
-                absmax = std::max(absmax, (float)fabs(bfloat16_to_float32(p0[0])));
+                absmax = std::max(absmax, (float)fabsf(bfloat16_to_float32(p0[0])));
                 p0 += A_hstep;
             }
 
@@ -2350,7 +2350,7 @@ static void compute_B_bf16_int8_scale(const Mat& B, float& scale)
 #endif
         for (; j < size; j++)
         {
-            absmax = std::max(absmax, (float)fabs(bfloat16_to_float32(ptr[0])));
+            absmax = std::max(absmax, (float)fabsf(bfloat16_to_float32(ptr[0])));
             ptr++;
         }
     }
